requirements.txt
datasets==2.17.1
einops
jsonlines
numexpr
openai>=0.6.4
omegaconf>=2.2
peft>=0.2.0
pybind11>=2.6.2
pycountry
pytablewriter
rouge-score>=0.0.4
sacrebleu==1.5.0
scikit-learn>=0.24.1
sqlitedict
tqdm-multiprocess
zstandard
accelerate
sentencepiece
tokenizers>=0.12.1
torch>=2.0.0
torchvision 
transformers==4.35.2 
texttable
toml
attributedict
protobuf